requirements.txt
fastapi>=0.100.0,<0.101.0
uvicorn>=0.23.0,<0.24.0
sqlalchemy>=1.4.41,<1.5.0
pydantic>=2.0.0,<3.0.0
python-jose>=3.3.0,<4.0.0
passlib>=1.7.4,<2.0.0
python-multipart>=0.0.5,<0.1.0
python-dotenv>=0.21.0,<0.22.0
alembic>=1.11.0,<1.12.0
bcrypt>=4.0.1,<5.0.0
requests>=2.31.0,<3.0.0